Celestino

+++ lib/binary_source.cpp
@@ -205,7 +205,7 @@ bool BinarySource::runBlock(void) {
 
 	if (mode == DeterministicCyclic){
 		std::vector<char> values(bitStream.begin(), bitStream.end());
-		int valuesSize = values.size();
+		int valuesSize = (int) values.size();
 		for (int k = 0; k < process; k++) {
 			t_binary aux = (t_binary)(values[posBitStream++] - '0');
 			for (auto k = 0; k < numberOfOutputSignals; k++) {
@@ -219,7 +219,7 @@ bool BinarySource::runBlock(void) {
 
 	if (mode == DeterministicAppendZeros){
 		std::vector<char> values(bitStream.begin(), bitStream.end());
-		int valuesSize = values.size();
+		int valuesSize = (int) values.size();
 		t_binary aux;
 		for (int k = 0; k < process; k++) {
 			if (posBitStream == valuesSize) {
